+++ src/Commands/CreateToken.php
@@ -58,6 +58,6 @@ class CreateToken extends Command
         $this->warn(sprintf('Information for the user %s', $userInformation));
         $this->info(sprintf('%s="%s"', app('protector')->getAuthTokenKeyName(), $token->plainTextToken));
         $this->warn('The quotation marks at the start and end of the token are necessary!');
-        $this->info(sprintf('%s=%s', app('protector')->getServerUrlKeyName(), route('protectorDumpEndpointRoute')));
+        $this->info(sprintf('%s=%s', app('protector')->getDumpEndpointUrlKeyName(), route('protectorDumpEndpointRoute')));
     }
 }